pyproject.toml
[project]
name = "goldener"
version = "0.1.0"
description = "Goldener - Make your data even more valuable"
readme = "README.md"
license = "Apache-2.0"
license-files= ["LICENSE"]
authors = [
    { name = "Yann-CV", email = "yann.chene.tni@gmail.com"}
]

requires-python = ">=3.10"
dependencies = [
    "torch",
    "opencv-python",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pre-commit",
  "ruff",
  "mypy",
  "types-pillow",
  "types-tqdm",
]

single_click = [
  "fiftyone",
  "pixeltable",
  "any-gold",
  "sam-2",
  "hydra-core",
  "huggingface-hub",
]

[tool.uv.sources]
any-gold = { git = "https://github.com/goldener-data/any-gold.git" }
sam-2 = { git = "https://github.com/facebookresearch/sam2.git" }
